build: add typescript config and update playwright setup with separate projects for v3 and c4
