+++ projects/conductor/package.json
@@ -11,9 +11,10 @@
   "devDependencies": {
     "@fingerprintjs/fingerprintjs-pro": "^3.11.6",
     "@fingerprintjs/fingerprintjs-pro-server-api": "6.10.0",
-    "@playwright/test": "^1.49.1",
+    "@playwright/test": "^1.58.2",
     "@types/node": "^22.10.1",
-    "dotenv": "^16.4.7"
+    "dotenv": "^16.4.7",
+    "typescript": "^5.6.3"
   },
   "dependencies": {
     "@fingerprint/node-sdk": "^7.0.0"
