+++ projects/conductor/tests/v3/004-searchEvents.test.ts
@@ -1,7 +1,7 @@
-import { test } from '../utils/playwright'
-import testData from '../utils/testData'
+import { test } from '../../utils/playwright'
+import testData from '../../utils/testData'
 import { expect } from '@playwright/test'
-import { delay } from '../utils/delay'
+import { delay } from '../../utils/delay'
 
 test.describe('SearchEvents suite', () => {
   test('with valid api key and limit', async ({ assert, identify }) => {
