+++ projects/conductor/tests/v4/001-getEvent.test.ts
@@ -1,7 +1,7 @@
-import { testData } from '../utils/testData'
-import { test } from '../utils/playwright'
+import { test } from '../../utils/v4/playwright'
+import testData from '../../utils/testData'
 
-test.describe('GetEvents Suite', () => {
+test.describe('GetEvent Suite', () => {
   test('for valid apiKey and requestId with Smart Signals', async ({ identify, assert }) => {
     const { requestId } = await identify({
       auth: testData.credentials.maxFeaturesUS,
@@ -9,7 +9,7 @@ test.describe('GetEvents Suite', () => {
     const requestData = {
       apiKey: testData.credentials.maxFeaturesUS.privateKey,
       region: testData.credentials.maxFeaturesUS.region,
-      requestId,
+      eventId: requestId,
     }
 
     await assert.thatResponsesMatch('getEvent', requestData)
@@ -22,7 +22,7 @@ test.describe('GetEvents Suite', () => {
     const requestData = {
       apiKey: testData.credentials.minFeaturesUS.privateKey,
       region: testData.credentials.minFeaturesUS.region,
-      requestId,
+      eventId: requestId,
     }
 
     await assert.thatResponsesMatch('getEvent', requestData)
@@ -35,17 +35,17 @@ test.describe('GetEvents Suite', () => {
         api.getEvent({
           apiKey: '',
           region: '',
-          requestId: '',
+          eventId: '',
         }),
     })
   })
 
-  test('for invalid apikey, region, and requestID', async ({ assert }) => {
+  test('for invalid apikey, region, and eventId', async ({ assert }) => {
     await assert.thatResponseMatch({
       expectedResponse: {
         error: {
-          code: 'TokenNotFound',
-          message: 'secret key is not found',
+          code: 'secret_api_key_not_found',
+          message: 'no fingerprint workspace found for specified secret API key',
         },
       },
       expectedStatusCode: 403,
@@ -53,7 +53,7 @@ test.describe('GetEvents Suite', () => {
         api.getEvent({
           apiKey: testData.credentials.invalid.privateKey,
           region: testData.credentials.invalid.region,
-          requestId: testData.invalid.requestID,
+          eventId: testData.invalid.requestID,
         }),
     })
   })
@@ -67,15 +67,15 @@ test.describe('GetEvents Suite', () => {
       expectedStatusCode: 403,
       expectedResponse: {
         error: {
-          code: 'TokenNotFound',
-          message: 'secret key is not found',
+          code: 'secret_api_key_not_found',
+          message: 'no fingerprint workspace found for specified secret API key',
         },
       },
       callback: (api) =>
         api.getEvent({
           apiKey: testData.credentials.invalid.privateKey,
           region: testData.credentials.maxFeaturesUS.region,
-          requestId,
+          eventId: requestId,
         }),
     })
   })
@@ -90,25 +90,25 @@ test.describe('GetEvents Suite', () => {
         api.getEvent({
           apiKey: testData.credentials.maxFeaturesUS.privateKey,
           region: testData.credentials.invalid.region,
-          requestId,
+          eventId: requestId,
         }),
     })
   })
 
-  test('for invalid requestId', async ({ assert }) => {
+  test('for invalid eventId', async ({ assert }) => {
     await assert.thatResponseMatch({
       expectedStatusCode: 404,
       expectedResponse: {
         error: {
-          code: 'RequestNotFound',
-          message: 'request id not found',
+          code: 'event_not_found',
+          message: 'event id not found',
         },
       },
       callback: (api) =>
         api.getEvent({
           apiKey: testData.credentials.maxFeaturesUS.privateKey,
           region: testData.credentials.maxFeaturesUS.region,
-          requestId: testData.invalid.requestID,
+          eventId: testData.invalid.requestID,
         }),
     })
   })
@@ -122,7 +122,7 @@ test.describe('GetEvents Suite', () => {
       expectedStatusCode: 403,
       expectedResponse: {
         error: {
-          code: 'WrongRegion',
+          code: 'wrong_region',
           message: 'wrong region',
         },
       },
@@ -130,7 +130,7 @@ test.describe('GetEvents Suite', () => {
         api.getEvent({
           apiKey: testData.credentials.maxFeaturesUS.privateKey,
           region: testData.credentials.regularEU.region,
-          requestId,
+          eventId: requestId,
         }),
     })
   })
@@ -144,15 +144,15 @@ test.describe('GetEvents Suite', () => {
       expectedStatusCode: 403,
       expectedResponse: {
         error: {
-          code: 'TokenNotFound',
-          message: 'secret key is not found',
+          code: 'secret_api_key_not_found',
+          message: 'no fingerprint workspace found for specified secret API key',
         },
       },
       callback: (api) =>
         api.getEvent({
           apiKey: testData.credentials.deleted.privateKey,
           region: testData.credentials.deleted.region,
-          requestId,
+          eventId: requestId,
         }),
     })
   })
